DataViews Extensibility: Allow unregistering permanently delete post action
