+++ packages/editor/src/dataviews/actions/index.ts
@@ -10,6 +10,7 @@ import deletePost from './delete-post';
 import exportPattern from './export-pattern';
 import resetPost from './reset-post';
 import trashPost from './trash-post';
+import permanentlyDeletePost from './permanently-delete-post';
 
 // @ts-ignore
 import { store as editorStore } from '../../store';
@@ -24,4 +25,5 @@ export default function registerDefaultActions() {
 	registerEntityAction( 'postType', '*', resetPost );
 	registerEntityAction( 'postType', '*', deletePost );
 	registerEntityAction( 'postType', '*', trashPost );
+	registerEntityAction( 'postType', '*', permanentlyDeletePost );
 }
